[TE] Added Stage::set_physical_axis_params.
In C++, this is implemented as an
`Array<BufferParamsPerPhysicalAxis>`, specifying where logical axes
should be flattened into physical axes, and details about those
physical axes.  The python interface uses a sentinel value
`te.PHYSICAL_AXIS_SEPARATOR` in the call to `set_physical_layout`,
which is then used to define the parameter sets.