fixup! Added pre_flattened_shape/pre_flattened_stride fields to Buffer.
Record pre-flattened shape/strides as Optional<Array<PrimExpr>>.

Previously, they were tracked as Array<PrimExpr>, with an empty array
indicating a buffer that hadn't yet been flattened.  This resulted in
ambiguity between a pre-StorageFlatten buffer of shape `[1]` and a
post-StorageFlatten scalar buffer, as both would have `shape=[1]` and
`pre_flatten_shape=[]`.

This ambiguity resulted in errors if StorageFlatten is called multiple
times on the same graph.