Test commit, making BufferNode hold pre_flattened_buffer.
Works, but requires copy-paste of the `TVM_DEFINE_OBJECT_REF_METHODS`
and `TVM_DEFINE_OBJECT_REF_COW_METHODS` macros, in order to separate
out the declaration and definitions.  Would like to have this or
something similar in the future, but will need to investigate
further.

Potentially holding the pre_flattened information inside the
`PrimFunc::buffer_map`, since that's the only place that needs it?