[Hexagon] Distinguish between 1-d buffer and single-alloc 2-d buffer
Previously, HexagonBuffer represented 1-d buffers as 2-d buffers with
`nallocs==1`.  Since this is used to determine the return type of the
data pointer exposed to the generated code, the ambiguity between
`shape=[N]` and `shape=[1,N]` must be avoided.  This commit replaces
`HexagonBuffer::nallocs_` with `HexagonBuffer::ndim_`, avoiding this
ambiguity.