[TIR] Reduce repeated walking in RewriteSimplifier
Prior to this commit, negating a expression required simplifying the
entire expression.  In many cases, such as setting up a constrained
scope for a conditional, numeric portions of the expression have
already been simplified, and do not need to be re-visited when taking
a negation.

This commit adds `tvm::arith::RewriteBooleanOperators`, which performs
the subset of rewrites necessary to negate an expression, but doesn't
visit numeric portions of the expression.