[TIR] Simplify chains of AndNode, chains of OrNode
Previously, application of pairwise simplification rules applied to a
nested call to `And(And(a, b), c)` would attempt to simplify `a` and
`b`, but would not attempt to simplify `a` and `c`.  As a result,
while `(i==j && i!=j)` would simplify to `false`, `(i==j && j==5) &&
i!=j` would not simplify.

After this PR, all conditions that contribute to an `And` are checked
for pairwise simplifications.