First attempt at double-dispatch for visitor, to be reverted.
Failed because virtual functions on `ObjectRef` subclasses are not
supported.  `ObjectRef` doesn't have a vtable, and `Downcast<T>`
constructs a new object of type `T` rather than casting to it.  As a
result, it is possible to have (for example) a `PrimExpr` instance
that was constructed as a `PrimExpr`, but holds a `CallNode`.  A
virtual function expecting to find a `Call` vtable would instead find
a `PrimExpr` vtable.