Next attempt, using a case/switch on a constexpr type id
Issue came up in making the type-id constexpr, which was required for
use in a case statement.  My usual trick of having a `void*` to a
static member of a templated type failed, because that isn't an
integer type, and `reinterpret_cast` isn't allowed in constexpr.  The
`std::type_info` isn't usable here, either, since none of its methods
are `constexpr`.

I think this will need to be reverted, and it will go back to
double-dispatch, but on the `Object` subclass instead of on the
`ObjectRef` subclass.