[TIR][Arith] Extract conditionals from ranges
This simplication applies only for integer comparisons against a
constant value.  The implementation separates out the independent
terms, then checks if incrementing or decrementing one of the
variables

For example, if iterators `i` and `j` are on ranges `0 <= i < 8` and
`0 <= j < 16`, then the condition `5*i + j < 4` can be split into
`i==0 && j<4`.

* Condition is only true when `i` uses its minimum allowed value.
  `5*i + j < 4` can be split into `i==0 && j<4`.

* Condition is only false when `i` uses its minimum allowed value.
  `4 < 5*i + j` can be split into `0<i || 4<j`.

* Condition is only true when `i` uses its maximum allowed value.
  `37 < 5*i + j` can be split into `i==7 && 2 < j`.

* Condition is only false when `i` uses its maximum allowed value.
  `5*i + j < 37` can be split into `i<7 || j < 2`.

This implementation does not use the utilities in
`iter_affine_map.h`, because the expression may not form an
injective transformation.  This is used for conditionals that
arise in conditionals for padded transformations, in operators
with a reduction axis.