[Relax] Avoid repeated variable bindings in BlockBuilder
Prior to this commit, using the `ExprMutator` to replace a value would
result in an extra variable assignment.  This was due to the
interaction of the `ExprMutator`, which would use a `SeqExpr` to
represent the modified expression, and `BlockBuilder`, which would
unwrap the nested `SeqExpr`.  This commit updates
`Normalize::FlattenBlocks` in `block_builder.cc` to avoid this
duplicate variable, when possible.