[TIR] Added CallNode::buffer_map
Similar to the `PrimFunc::buffer_map` used at a callee's definition,
the `CallNode::buffer_map` describes buffers that should be exposed
to a subroutine.  Unlike the `PrimFunc::buffer_map`, which is a
`Map<Var, Buffer>`, the `CallNode::buffer_map` is a `Map<Var,
BufferRegion>`, as a subroutine may not receive the entire buffer as
an input.