[TIR] Implement LowerBufferArguments (requires CallNode::buffer_map)
Prior to this commit, only primitives could be used as arguments in
`Call` nodes.  This commit allows buffer objects in the
`CallNode::buffer_map` to be written as arguments by the user.  These
buffer arguments are then lowered into either primitives for internal
subroutines, or to `DLTensor` arguments for subroutines that are
exposed externally.