[TIR] Handle implicit buffers in PrimFunc::Specialize
Prior to this commit, the `tvm::tir::Specialize` function would
replace the specialized parameter when it occurs in the shape,
strides, or element offset of a buffer, but only when the buffer
occurs in specific locations.  As a result, a `tir::Buffer` in the
`PrimFuncNode::buffer_map` and `BlockNode` would be updated, but
`tir::Buffer` objects in a `BufferLoad` or `BufferStore` node could be
left unmodified and therefore contain undefined variables.

This commit updates `PrimFunc::Specialize` to update buffer objects in
`BufferLoad` and `BufferStore`.