[TIR][Buffer] Implement utility method AsDLTensor for Buffer/Region
Prior to this commit, `tvm::topi::detail::pack_buffer()` utility
function could be used to exposes a `tir::Buffer` as a `DLTensor`,
returning a `PrimExpr` that is suitable for use as an argument to
external function calls that accept a `DLTensor*`.  This commit moves
this functionality to methods of `tir::Buffer::AsDLTensor` and
`tir::BufferRegion::AsDLTensor`, with the existing `pack_buffer()`
function retained for backwards compatibility.