[TIR] Retain kIsHostFunc, previous kTarget, in BindTarget
Prior to this commit, the `tir.BindTarget` pass would strip out any
`tvm::tir::attr::kIsHostFunc` attributes (`"tir.is_host_func"`), and
would replace any prior `tvm::attr::kTarget` attributes (`"target"`).
This commit updates the behavior so that any previously defined
attributes for `tvm::attr::kTarget` are kept unmodified, and the
`tvm::tir::attr::kIsHostFunc` attribute is retained in the output.

This is part of a series of commits to simplify the handling of
multi-target builds.