[TIR] Apply the kIsHostFunc attribute to SplitHostDevice output
Previously, the host function generated by `tir.SplitHostDevice` was
stripped of any `tvm::attr::kTarget` annotation.  This commit updates
the `tir.SplitHostDevice` pass so that it replaces the
`tvm::attr::kTarget` annotation with the host, and also inserts a
`tvm::tir::attr::kIsHostFunc` attribute.

This is part of a series of commits to simplify the handling of
multi-target builds.  By tracking the host functions separately,
passes that previously applied to just the host `IRModule` can be
applied to the merged host/device `IRModule`, while only altering the
host functions.