[TIR] Check kIsHostFunc in CombineContextCall and BuiltinLower
These lowering passes should only alter functions that occur on the
host side after `tir.SplitHostDevice`.

This is part of a series of commits to simplify the handling of
multi-target builds.  By ignoring device-side functions, these passes
can be applied to a mixed `IRModule` containing both host- and
device-side functions.