[TIR] Retain device id/type attributes after LowerTVMBuiltin
Previously, the `tir.LowerTVMBuiltin` pass would strip out the
`tvm::tir::attr::device_id` and `tvm::tir::attr::device_type`
annotations after reading them.  As a result, if `LowerTVMBuiltin`
would use a stack-based allocation for a `tir::Allocate` node on a
target that runs on `kDLCPU`, the resulting TIR would appear to be
missing the device id/type attributes.

This commit updates `tir.LowerTVMBuiltin` to preserve the device
id/type attributes, so that the lowering pass can be idempotent.