[TVMScript] Provide default name for dynamic T.Buffer parameters
Prior to this commit, the name used for a buffer argument was only
applied to the buffer itself, and not to any otherwise-unnamed
variables in the buffer.  (e.g. The buffer shape in a PrimFunc with
argument `A: T.Buffer(shape=[T.int32(), T.int32()])`.)  Variables with
an empty name are legal in TIR, but can cause issues for code
generators derived from `CodeGenC` (e.g. `CodeGenCUDA`), which use the
TIR variable name as the C variable name.  This can cause compilation
failure, as an empty variable name results in syntax errors in the
generated code.

This commit applies a default name for variables in TIR buffers, if
they would otherwise have an empty name.