[TVMScript] Avoid renaming named variables in buffer
Prior to this commit, naming a buffer would rename all stride
variables in the buffer definition, even if they were already named.

This commit updates the behavior to only rename variables that are
currently named `"v"`, the default name in the C++ constructor, and
the empty string `""`.  All other variable names are kept as-is.