[TVMScript] Validate sign of literal indices
Neither TVMScript nor TIR support negative indices.  Since TVMScript
strongly resembles Python's syntax, a user might assume that negative
indices will wrap around to the end of a buffer slice.

While this could be supported solely on the TVMScript side for integer
literals, it wouldn't be able to handle cases where a TIR variable has
been assigned a negative value, and is then used as an index.
Therefore, this commit instead checks for the most common case of an
integer literal and provides a warning.