[Unity] Handle bound tuple variables in relax.transform.FoldConstant
Prior to this commit, the `relax.transform.FoldConstant` pass assumed
that the `""relax.call_tir"` builtin had arguments expressed as
`relax.Tuple`, and failed if provided with a `relax.Var` that had been
bound to a tuple.  This commit updates the `FoldConstant` pass to
handle variables annotated with `TupleStructInfo`.

If the variable's value was determined within the scope of the mutated
function, we can look up the bound tuple and find the argument.  If
the variable's value was produced as output from another function,
then we cannot use it in a constant expression, and must leave it
as-is.