[Unity] Handle tuple variables in relax.transform.FuseOps
Prior to this commit, the `relax.transform.FuseOps` pass assumed that
the `""relax.call_tir"` builtin had arguments expressed as
`relax.Tuple`, and failed if provided with a `relax.Var` that had been
bound to a tuple.  This commit updates the `FuseOps` pass to unwrap
variable bindings prior to downcasting from `relax::Expr` to
`relax::Tuple`.