[Unity] Handle tuple variables in RewriteDataflowReshape
Prior to this commit, the `relax.transform.RewriteDataflowReshape`
pass assumed that the `""relax.call_tir"` builtin had arguments
expressed as `relax.Tuple`, and failed if provided with a `relax.Var`
that had been bound to a tuple.  This commit updates the
`RewriteDataflowReshape` pass to handle variables annotated with
`TupleStructInfo`.

The identification of a reshape can be done with only the number of
arguments, which can be extracted from the variables `TupleStructInfo`
instead of requiring a `Tuple`.  If the TIR function is a reshape,
then the tuple variable can be unwrapped to a known tuple in order to
find the argument, or can use a `TupleGetItem` node to extract the
argument.