[Unity][Transform] Check for infer-able expressions in FuseOps
Prior to this commit, `FuseOps` and `FuseOpsByPattern` exposed
a symbolic variable to the fused function if it was used within the
fused function, but wasn't inferable from other parameter shapes.
While this prevents undefined symbolic variables, it can cause issues
for downstream use of `CodegenJSON`, which requires all arguments to
be tensors, or tuple of tensors.

Frequently, all uses of a non-inferable symbolic shape occur within a
symbolic expression that can be inferred.  For example, a function
that takes `arg: R.Tensor([N+1])` and returns `R.add(arg, R.const(1))`
cannot infer `N`.  However, all occurrences of `N` occur as part
of the expression `N+1`, and the value of `N+1` can be inferred.
Therefore, if we replace `N+1` with `M`, the additional `ShapeTuple`
argument isn't required.