[Relax] Ignore TIR PrimFunc in relax DeadCodeElimination
Prior to this commit, an externally-exposed PrimFunc would be identified
as an entry function in `relax.transform.DeadCodeElimination`,
resulting in an error when attempting to downcast into
`relax.Function`.  This commit updates `DeadCodeElimination` to skip
visits into a `PrimFunc`.

For the long-term, because a `PrimFunc` could be called either from a
`relax::Function` or from another `PrimFunc`, the elimination of
uncalled functions should be moved to a common transform that can
track callsites in both TIR and Relax functions.