Handle relax.Const in LazyTransformParams
Prior to this commit, `LazyTransformParams` would only output a call
to the `fset_item` function if that element of the output had a
corresponding `relax.Binding`.  If `relax.Const` appeared in the
output, then the call to `fset_item` would be omitted.

This commit updates `LazyTransformParams` to check for any non-`Var`
elements of the output tuple.