[QoL][Relax] Infer StructInfo for relax::SeqExpr on construction
Prior to this commit, the `relax::SeqExpr` constructor left the
`struct_info_` field undefined.  This is inconsistent with other Relax
leaf nodes, such as `relax::PrimValue`, `relax::Constant`, and
`relax::ExternFunc`, which initialize their struct info on
construction.

This commit updates the `relax::Tuple` constructor to define
`struct_info_` as `TupleStructInfo`, if all fields have a known struct
info.  If any field does not have a known struct info, the current
behavior is kept, where `struct_info_` is constructed as `NullOpt`,
and is later populated by the `relax::BlockBuilder`.

There are cases in which the `relax::SeqExpr` should have `StructInfo`
distinct from its body.  For example, the `StructInfo` of the body may
reference a symbolic variable that is defined by a `relax::MatchCast`
within the bindings of the `SeqExpr`.  However, the IR representations
of a `relax::MatchCast` that defines a symbolic variable and a
`relax::MatchCast` that uses a previously-defined symbolic variable
are identical.  Distinguishing between these cases requires contextual
information, knowing which symbolic variables are defined prior to the
`SeqExpr`.  This is handled in `relax::Normalizer::VisitWithNewScope`, which
has access to both the scope of the `SeqExpr` and the scope containing
the `SeqExpr`.

TODO: Handle those cases, using CopyOnWrite in VisitWithNewScope if
there are any symbolic variables that need to be erased.